Fixes #26. Fixed dateFromSQLTimestamp function.
git-svn-id: http://www.phpbms.org/svn/trunk@115 66a7d52b-23fa-0310-b706-d52274950f5d

+++ phpbms/include/common_functions.php
@@ -148,7 +148,7 @@ function dateFromSQLTimestamp ($datetime) {
 	array_shift ($matches);	
 	foreach (array('year','month','day','hour','minute','second') as
 $var) {
-		$$var = array_shift($matches);
+		$$var = (int) array_shift($matches);
 	}
 	return mktime($hour,$minute,$second,$month,$day,$year);
 }
